% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diameter_correction.R
\name{correct_diameter_single}
\alias{correct_diameter_single}
\title{Correct diameter values of an artery from a single patient dataset.}
\usage{
correct_diameter_single(
  df_patient,
  sdUS = 2,
  sdCT = 1,
  sp = 4,
  dlim_sup = 1000,
  dlim_inf = 0
)
}
\arguments{
\item{df_patient}{A data.frame with columns Date (numeric), Diam, CT.}

\item{sdUS}{Standard deviation of the diameter measured by Ultrasound, in the same units as the dataframe Diam values. 2 as default.}

\item{sdCT}{Standard deviation of the diameter measured by CT-scan in the same units as the dataframe Diam values. 1 as default.}

\item{sp}{Number of points to discretize the probability distribution. 4 as default.}

\item{dlim_sup}{Maximum allowed growth (Diam2-Diam1)/(Date2-Date1). 50 as default.}

\item{dlim_inf}{Minimum allowed growth (Diam2-Diam1)/(Date2-Date1). 0 as default.}
}
\value{
A list with 3 elements. \code{curves} vector that contains the curves with the maximum probability of existence that fits the input conditions. \code{max_prob} The product of the probability of each point of the curve. \code{num_curves} number of curves with the maximum probability.
}
\description{
Takes into account the error of the diameter measures from an artery depending the technique that have been used for measuring it.
Generates additional diameter measures with an associated probability of being real based on the standard deviation of the measure.
Finds the most probable diameter progression curve using Viterbi algorithm.
}
\examples{
data <- data.frame(
 ID = as.factor(c("1", "1", "1", "1", "1", "1", "1", "1", "1", "2", "2", "2", "2")),
 Date = c(2015, 2016, 2017, 2020, 2021, 2024, 2025, 2026, 2027, 2013, 2016, 2018, 2019),
 Diam = c(30, 35, 34, 53, 50, 52, 55, 55, 57, 38, 42, 50, 53),
 CT = as.factor(c(1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1))
)

# Filter patients
df_patient <- data[ID==1,]
res <- correct_diameter_single(df_patient)
}
