% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diameter_correction.R
\name{correct_diameters_all}
\alias{correct_diameters_all}
\title{Correct diameter values of an artery from a multi-patient dataset.}
\usage{
correct_diameters_all(
  data,
  sdUS = 2,
  sdCT = 1,
  sp = 4,
  dlim_inf = 0,
  dlim_sup = 50
)
}
\arguments{
\item{sdUS}{Standard deviation of the diameter measured by Ultrasound, in the same units as the dataframe Diam values. 2 as default.}

\item{sdCT}{Standard deviation of the diameter measured by CT-scan in the same units as the dataframe Diam values. 1 as default.}

\item{sp}{Number of points to discretize the probability distribution. 4 as default.}

\item{dlim_inf}{Minimum allowed growth (Diam2-Diam1)/(Date2-Date1). 0 as default.}

\item{dlim_sup}{Maximum allowed growth (Diam2-Diam1)/(Date2-Date1). 50 as default.}

\item{df_patient}{A data.frame with columns ID, Date, Diam, CT.}
}
\value{
A dataframe with the same number of rows as the input dataframe with corrected diameter values and a additional column that specifies if the measure have been corrected or not.
}
\description{
Takes into account the error of the diameter measures from an artery depending the technique that have been used for measuring it.
Generates additional diameter measures with an associated probability of being real based on the standard deviation of the measure.
Finds the most probable diameter progression curve using Viterbi algorithm and returns a dataframe with the corrected values.
The output dataset has an additional column where it specifies if the diameter progression curve have been corrected or not.
}
\examples{
data <- data.frame(
 ID = as.factor(c("1", "1", "1", "1", "1", "1", "1", "1", "1", "2", "2", "2", "2")),
 Date = c(2015, 2016, 2017, 2020, 2021, 2024, 2025, 2026, 2027, 2013, 2016, 2018, 2019),
 Diam = c(30, 35, 34, 53, 50, 52, 55, 55, 57, 38, 42, 50, 53),
 CT = as.factor(c(1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1))
)

res <- correct_diameter_all(df_patient)
}
