% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a6er_utils.R
\name{MAD_precision}
\alias{MAD_precision}
\title{Mean Absolute Difference (MAD) precision calculation}
\usage{
MAD_precision(df)
}
\arguments{
\item{df}{Dataframe containing the mean (mn) and difference (diff) columns}
}
\value{
The MAD Precision value in \%.
}
\description{
Calculates the mean absolute difference precision of a dataframe that contains the
mean and difference values of measured duplicates.
}
\examples{
MAD1 <- MAD_precision(data.frame(mn = c(2 , 5 , 6), diff(0.1, 0.5, 0.2)));
}
