% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a6er_utils.R
\name{mean_pairwise_diff}
\alias{mean_pairwise_diff}
\title{Mean pairwise calculation}
\usage{
mean_pairwise_diff(values)
}
\arguments{
\item{values}{The measurement duplicated values of the same sample.}
}
\value{
The difference of the measurements divided by sqrt(2) and the number of samples used for the calculus.
}
\description{
Calculates the arithmetic differences between duplicate measurement pairs.
If the same measure have been measured more than twice, it computes the mean of the differences.
The differences are divided by sqrt(2) because they arise from imprecision in both
measurements, which adds quadratically, and the desired result is the precision
of one measurement.
}
\examples{
diff1 <- mean_pairwise_diff(c(22, 26));
diff2 <- mean_pairwise_diff( c(10, 12, 15) );
}
