% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precission_error_calculation.R
\name{measure_precision}
\alias{measure_precision}
\title{Measurement Precision calculation}
\usage{
measure_precision(tb, img = TRUE)
}
\arguments{
\item{tb}{A dataframe with \code{id} (unique ID value per duplicated sample), \code{sensor} (type of sensors) and \code{measurements} (the results of the measurements) columns}

\item{img}{(LOGICAL element) It determines if you wants plots or not. It is TRUE as default.}
}
\value{
A tibble grouped by sensor that contains a dataframe with the next three columns 1.difference between duplicates, 2.mean values of each duplicate and 3.number of samples of the duplicates (commonly there will be just 2 sample as the name duplicate says, but if there are triplicates or more, the program will work as well is why it is important to know the n of samples) and RMS, MAD precision values and the number of samples that have been used to calculate these values (n_precision).
}
\description{
Calculates the arithmetic differences between duplicate measurement pairs.
If the same measure have been measured more than twice, it computes the mean of the differences.
The differences are divided by sqrt(2) because they arise from imprecision in both
measurement, which adds quadratically, and the desired result is the precision
of one measurement.
}
\examples{
tb <- data.frame(id = c("m1", "m1", "m1", "m1", "m2", "m2", "m2", "m2"),
                 sensor = c("s1", "s1", "s2", "s2", "s1", "s1", "s2", "s2"),
                 measurement = c(10, 9, 2, 1, 15, 17, 4, 3))
tb_precision <- measure_precision(tb);
}
