% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a6er_utils.R
\name{RMS_precision}
\alias{RMS_precision}
\title{Root Mean Square precision calculation}
\usage{
RMS_precision(df)
}
\arguments{
\item{df}{Dataframe containing the mean (mn) and difference (diff) columns}
}
\value{
The Root Mean Square Precision value in \%.
}
\description{
Calculates the root mean square precision of a dataframe that contains the
mean and difference values of measured duplicates.
}
\examples{
RMS1 <- RMS_precision(data.frame(mn = c(2 , 5 , 6), diff(0.1, 0.5, 0.2)));
}
